reorganized code to separate join interface from join algorithm
