+++ core/src/main/java/tech/tablesaw/joining/JoinType.java
@@ -1,0 +1,9 @@
+package tech.tablesaw.joining;
+
+/** The types of joins that are supported */
+enum JoinType {
+  INNER,
+  LEFT_OUTER,
+  RIGHT_OUTER,
+  FULL_OUTER
+}
